requirements.txt
numpy>=2.2.5
pandas>=2.2.3
scikit-learn>=1.6.1
torch>=2.6.0
matplotlib>=3.10.1
seaborn>=0.13.2
pyyaml>=6.0.1
pytest>=8.3.5
psutil>=7.0.0
dask>=2025.4.1
fastapi>=0.115.0
uvicorn>=0.31.0
pyarrow>=20.0.0
pydantic>=2.11.0
tqdm>=4.66.1
joblib>=1.3.2